--- .eslintrc.json
@@ -1,42 +1,0 @@
-{
-  "extends": [
-    "airbnb",
-    "airbnb-typescript"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "./tsconfig.json"
-  },
-  "rules": {
-    "react/function-component-definition": "off",
-    "import/prefer-default-export": "off",
-    "react/react-in-jsx-scope": "off",
-    "import/extensions": "off",
-    "jsx-a11y/control-has-associated-label": "off",
-    "import/order": [
-      "error",
-      {
-        "pathGroups": [
-          { "pattern": "react", "group": "builtin" },
-          { "pattern": "vite", "group": "builtin" },
-          { "pattern": "~shared/**", "group": "internal" },
-          { "pattern": "~entities/**", "group": "internal" },
-          { "pattern": "~features/**", "group": "internal" },
-          { "pattern": "~widgets/**", "group": "internal" },
-          { "pattern": "~pages/**", "group": "internal" }
-        ],
-        "pathGroupsExcludedImportTypes": ["builtin"],
-        "groups": [
-          "builtin",
-          "external",
-          "internal",
-          "parent",
-          "sibling",
-          "index"
-        ],
-        "newlines-between": "never",
-        "alphabetize": { "order": "asc", "caseInsensitive": true }
-      }
-    ]
-  }
-}
