+++ package.json
@@ -1,46 +1,37 @@
 {
-  "name": "skpro",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "eslint .",
-    "lint:fix": "eslint --fix --ext .js,.jsx,.ts,tsx .",
-    "prepare": "husky"
-  },
-  "dependencies": {
-    "@types/node": "20.2.5",
-    "@types/react": "18.2.7",
-    "@types/react-dom": "18.2.4",
-    "autoprefixer": "10.4.14",
-    "classnames": "^2.3.2",
-    "eslint-config-next": "13.4.4",
-    "fslightbox-react": "^1.7.4",
-    "next": "13.4.4",
-    "postcss": "8.4.24",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-icons": "^4.9.0",
-    "react-redux": "^8.0.7",
-    "react-responsive": "^9.0.2",
-    "sass": "^1.62.1",
-    "sharp": "^0.33.3",
-    "tailwindcss": "3.3.2",
-    "typescript": "5.0.4"
-  },
-  "devDependencies": {
-    "@types/fslightbox-react": "^1.7.3",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "eslint": "^8.2.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "husky": "^9.0.11"
-  }
+	"name": "skpro",
+	"version": "2.0.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"prepare": "husky",
+		"check": "biome check . --write"
+	},
+	"dependencies": {
+		"clsx": "^2.1.1",
+		"fslightbox-react": "^2.0.3",
+		"lucide-react": "^0.575.0",
+		"next": "16.1.6",
+		"radix-ui": "^1.4.3",
+		"react": "19.2.4",
+		"react-dom": "19.2.4",
+		"react-icons": "^5.5.0",
+		"react-responsive": "^10.0.1"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "^2.4.4",
+		"@tailwindcss/postcss": "^4",
+		"@types/fslightbox-react": "^2.0.1",
+		"@types/node": "^25",
+		"@types/react": "^19",
+		"@types/react-dom": "^19",
+		"husky": "^9.1.7",
+		"tailwindcss": "^4",
+		"typescript": "^5"
+	},
+	"volta": {
+		"node": "24.12.0"
+	}
 }
