--- src/app/components/Pagination/pagination.scss
@@ -1,86 +1,0 @@
-@tailwind components;
-
-@layer components {
-  .bg-grey {
-    background: rgb(214, 219, 220);
-  }
-}
-
-.pagination-container {
-  display: flex;
-  list-style-type: none;
-
-  .pagination-item {
-    padding: 0 12px;
-    height: 32px;
-    text-align: center;
-    margin: auto 4px;
-    @apply text-black dark:text-white;
-    display: flex;
-    box-sizing: border-box;
-    align-items: center;
-    letter-spacing: 0.01071em;
-    border-radius: 16px;
-    line-height: 1.43;
-    font-size: 13px;
-    min-width: 32px;
-
-    &.dots:hover {
-      background-color: transparent;
-      cursor: default;
-      @apply text-black;
-    }
-
-    &:hover {
-      @apply bg-grey text-black;
-      cursor: pointer;
-    }
-
-    &:hover .arrow::before {
-      @apply border-black dark:border-black;
-    }
-
-    &.selected {
-      @apply bg-grey text-black;
-    }
-
-    .arrow {
-      &::before {
-        position: relative;
-        /* top: 3pt; Uncomment this to lower the icons as requested in comments*/
-        content: '';
-        /* By using an em scale, the arrows will size with the font */
-        display: inline-block;
-        width: 0.4em;
-        height: 0.4em;
-        border-right: 0.12em solid;
-        border-top: 0.12em solid;
-        @apply border-black dark:border-white;
-      }
-
-      &.left {
-        transform: rotate(-135deg) translate(-50%);
-      }
-
-      &.right {
-        transform: rotate(45deg);
-      }
-    }
-
-    &.disabled {
-      display: none;
-      pointer-events: none;
-
-      .arrow::before {
-        border-right: 0.12em solid;
-        border-top: 0.12em solid;
-        @apply border-black dark:border-white;
-      }
-
-      &:hover {
-        background-color: transparent;
-        cursor: default;
-      }
-    }
-  }
-}
