--- src/app/globals.css
@@ -1,67 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-* {
-    transition: 80ms;
-}
-
-@layer components {
-    .bg-gradient-white {
-        background: linear-gradient(to bottom, transparent, rgb(255, 255, 255)), rgb(214, 219, 220);
-    }
-
-    .bg-gradient-black {
-        background: linear-gradient(to bottom, transparent, rgb(40, 40, 40)), rgb(0, 0, 0);
-    }
-
-    .bg-grey {
-        background: rgb(214, 219, 220);
-    }
-
-}
-
-@layer utilities {
-    /* Hide scrollbar for Chrome, Safari and Opera */
-    .no-scrollbar::-webkit-scrollbar {
-        display: none;
-    }
-    /* Hide scrollbar for IE, Edge and Firefox */
-    .no-scrollbar {
-        -ms-overflow-style: none;  /* IE and Edge */
-        scrollbar-width: none;  /* Firefox */
-    }
-}
-
-.fp-overflow {
-    overflow: hidden !important;
-    -ms-overflow-style: none;
-    scrollbar-width: none;
-}
-
-.fp-overflow::-webkit-scrollbar {
-    width: 0;
-    height: 0;
-}
-
-body {
-    @apply text-black dark:text-white bg-gradient-white dark:bg-gradient-black;
-}
-
-
-.fp-watermark {
-    display: none;
-}
-
-.fullpage-wrapper {
-    width: 100%;
-}
-
-.fslightbox-open{
-    height: auto !important;
-}
-
-
-#fp-nav ul li a span {
-    @apply bg-black dark:bg-white;
-}
