--- src/app/page.module.scss
@@ -1,15 +1,0 @@
-.container {
-  position: relative;
-  overflow-y: auto;
-  overscroll-behavior-y: contain;
-  scroll-snap-type: y mandatory;
-  min-height: 100vh;
-  height: 100vh;
-  scroll-behavior: smooth;
-  width: 100%;
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
