--- src/app/projects/ProjectData.ts
@@ -1,402 +1,0 @@
-import { StaticImageData } from 'next/image';
-import chd1 from '../../../public/projects/chd/chd1.png';
-import chd2 from '../../../public/projects/chd/chd2.png';
-import chd3 from '../../../public/projects/chd/chd3.png';
-import chd4 from '../../../public/projects/chd/chd4.png';
-import dolinacapital1 from '../../../public/projects/dolinacapital/dolinacapital1.png';
-import dolinacapital2 from '../../../public/projects/dolinacapital/dolinacapital2.png';
-import dolinacapital3 from '../../../public/projects/dolinacapital/dolinacapital3.png';
-import dolinacapital4 from '../../../public/projects/dolinacapital/dolinacapital4.png';
-import edocs1 from '../../../public/projects/edocs/edocs1.png';
-import edocs2 from '../../../public/projects/edocs/edocs2.png';
-import edocs3 from '../../../public/projects/edocs/edocs3.png';
-import edocs4 from '../../../public/projects/edocs/edocs4.png';
-import edocs5 from '../../../public/projects/edocs/edocs5.png';
-import finAlley1 from '../../../public/projects/finAlley/finAlley1.png';
-import finAlley10 from '../../../public/projects/finAlley/finAlley10.png';
-import finAlley11 from '../../../public/projects/finAlley/finAlley11.png';
-import finAlley12 from '../../../public/projects/finAlley/finAlley12.png';
-import finAlley2 from '../../../public/projects/finAlley/finAlley2.png';
-import finAlley3 from '../../../public/projects/finAlley/finAlley3.png';
-import finAlley4 from '../../../public/projects/finAlley/finAlley4.png';
-import finAlley5 from '../../../public/projects/finAlley/finAlley5.png';
-import finAlley6 from '../../../public/projects/finAlley/finAlley6.png';
-import finAlley7 from '../../../public/projects/finAlley/finAlley7.png';
-import finAlley8 from '../../../public/projects/finAlley/finAlley8.png';
-import finAlley9 from '../../../public/projects/finAlley/finAlley9.png';
-import itAcademy1 from '../../../public/projects/itAcademy/itAcademy1.png';
-import itAcademy10 from '../../../public/projects/itAcademy/itAcademy10.png';
-import itAcademy11 from '../../../public/projects/itAcademy/itAcademy11.png';
-import itAcademy12 from '../../../public/projects/itAcademy/itAcademy12.png';
-import itAcademy2 from '../../../public/projects/itAcademy/itAcademy2.png';
-import itAcademy3 from '../../../public/projects/itAcademy/itAcademy3.png';
-import itAcademy4 from '../../../public/projects/itAcademy/itAcademy4.png';
-import itAcademy5 from '../../../public/projects/itAcademy/itAcademy5.png';
-import itAcademy6 from '../../../public/projects/itAcademy/itAcademy6.png';
-import itAcademy7 from '../../../public/projects/itAcademy/itAcademy7.png';
-import itAcademy8 from '../../../public/projects/itAcademy/itAcademy8.png';
-import itAcademy9 from '../../../public/projects/itAcademy/itAcademy9.png';
-import kadabra1 from '../../../public/projects/kadabra/kadabra1.png';
-import kadabra2 from '../../../public/projects/kadabra/kadabra2.png';
-import kadabra3 from '../../../public/projects/kadabra/kadabra3.png';
-import kadabra4 from '../../../public/projects/kadabra/kadabra4.png';
-import kadabra5 from '../../../public/projects/kadabra/kadabra5.png';
-import kadabra6 from '../../../public/projects/kadabra/kadabra6.png';
-import kadabra7 from '../../../public/projects/kadabra/kadabra7.png';
-import kadabra8 from '../../../public/projects/kadabra/kadabra8.png';
-import kadabra9 from '../../../public/projects/kadabra/kadabra9.png';
-import lagom1 from '../../../public/projects/lagom/lagom1.png';
-import lagom2 from '../../../public/projects/lagom/lagom2.png';
-import lagom3 from '../../../public/projects/lagom/lagom3.png';
-import lagom4 from '../../../public/projects/lagom/lagom4.png';
-import mate1 from '../../../public/projects/mate/mate1.png';
-import mate10 from '../../../public/projects/mate/mate10.png';
-import mate11 from '../../../public/projects/mate/mate11.png';
-import mate12 from '../../../public/projects/mate/mate12.png';
-import mate13 from '../../../public/projects/mate/mate13.png';
-import mate14 from '../../../public/projects/mate/mate14.png';
-import mate15 from '../../../public/projects/mate/mate15.png';
-import mate16 from '../../../public/projects/mate/mate16.png';
-import mate17 from '../../../public/projects/mate/mate17.png';
-import mate2 from '../../../public/projects/mate/mate2.png';
-import mate3 from '../../../public/projects/mate/mate3.png';
-import mate4 from '../../../public/projects/mate/mate4.png';
-import mate5 from '../../../public/projects/mate/mate5.png';
-import mate6 from '../../../public/projects/mate/mate6.png';
-import mate7 from '../../../public/projects/mate/mate7.png';
-import mate8 from '../../../public/projects/mate/mate8.png';
-import mate9 from '../../../public/projects/mate/mate9.png';
-import rnd1 from '../../../public/projects/rnd/rnd1.png';
-import rnd10 from '../../../public/projects/rnd/rnd10.png';
-import rnd11 from '../../../public/projects/rnd/rnd11.png';
-import rnd12 from '../../../public/projects/rnd/rnd12.png';
-import rnd13 from '../../../public/projects/rnd/rnd13.png';
-import rnd14 from '../../../public/projects/rnd/rnd14.png';
-import rnd2 from '../../../public/projects/rnd/rnd2.png';
-import rnd3 from '../../../public/projects/rnd/rnd3.png';
-import rnd4 from '../../../public/projects/rnd/rnd4.png';
-import rnd5 from '../../../public/projects/rnd/rnd5.png';
-import rnd6 from '../../../public/projects/rnd/rnd6.png';
-import rnd7 from '../../../public/projects/rnd/rnd7.png';
-import rnd8 from '../../../public/projects/rnd/rnd8.png';
-import rnd9 from '../../../public/projects/rnd/rnd9.png';
-import skylab1 from '../../../public/projects/skylab/skylab1.png';
-import skylab10 from '../../../public/projects/skylab/skylab10.png';
-import skylab11 from '../../../public/projects/skylab/skylab11.png';
-import skylab2 from '../../../public/projects/skylab/skylab2.png';
-import skylab3 from '../../../public/projects/skylab/skylab3.png';
-import skylab4 from '../../../public/projects/skylab/skylab4.png';
-import skylab5 from '../../../public/projects/skylab/skylab5.png';
-import skylab6 from '../../../public/projects/skylab/skylab6.png';
-import skylab7 from '../../../public/projects/skylab/skylab7.png';
-import skylab8 from '../../../public/projects/skylab/skylab8.png';
-import skylab9 from '../../../public/projects/skylab/skylab9.png';
-import techyard1 from '../../../public/projects/techyard/techyard1.png';
-import techyard2 from '../../../public/projects/techyard/techyard2.png';
-import techyard3 from '../../../public/projects/techyard/techyard3.png';
-import techyard4 from '../../../public/projects/techyard/techyard4.png';
-import truckstaff1 from '../../../public/projects/truckstaff/truckstaff1.png';
-import truckstaff2 from '../../../public/projects/truckstaff/truckstaff2.png';
-import undp1 from '../../../public/projects/undp/undp1.png';
-import undp10 from '../../../public/projects/undp/undp10.png';
-import undp11 from '../../../public/projects/undp/undp11.png';
-import undp12 from '../../../public/projects/undp/undp12.png';
-import undp13 from '../../../public/projects/undp/undp13.png';
-import undp2 from '../../../public/projects/undp/undp2.png';
-import undp3 from '../../../public/projects/undp/undp3.png';
-import undp4 from '../../../public/projects/undp/undp4.png';
-import undp5 from '../../../public/projects/undp/undp5.png';
-import undp6 from '../../../public/projects/undp/undp6.png';
-import undp7 from '../../../public/projects/undp/undp7.png';
-import undp8 from '../../../public/projects/undp/undp8.png';
-import undp9 from '../../../public/projects/undp/undp9.png';
-import valstatex1 from '../../../public/projects/valstatex/valstatex1.png';
-import valstatex2 from '../../../public/projects/valstatex/valstatex2.png';
-import valstatex3 from '../../../public/projects/valstatex/valstatex3.png';
-
-export type ProjectType = {
-  id: string
-  name: string
-  shortDescription: string
-  logo: string
-  description: string
-  link: string
-  stack: string
-  photos: PhotosType[]
-};
-
-type PhotosType = {
-  id: number
-  link: StaticImageData
-};
-
-export const ProjectsData: ProjectType[] = [
-  {
-    id: '1',
-    name: 'Lagom.software',
-    shortDescription: 'Аутсорсинговая компания Lagom.software',
-    logo: '/projects/lagom/lagom.svg',
-    description: 'Сайт портфолио с услугами аутсорсинговой компании.',
-    link: '',
-    stack: 'ReactJS, Ant Design, i18next',
-    photos: [
-      { id: 1, link: lagom1 },
-      { id: 2, link: lagom2 },
-      { id: 3, link: lagom3 },
-      { id: 4, link: lagom4 },
-    ],
-  },
-  {
-    id: '2',
-    name: 'Dolina Capital',
-    shortDescription: 'Портфолио компании инвесторов Dolina Capital',
-    logo: '/projects/dolinacapital/dolinacapital.svg',
-    description: 'Портфолио инвесторов где можно посмотреть их улсуги и их проекты.',
-    link: '',
-    stack: 'ReactJS, Ant Design',
-    photos: [
-      { id: 1, link: dolinacapital1 },
-      { id: 2, link: dolinacapital2 },
-      { id: 3, link: dolinacapital3 },
-      { id: 4, link: dolinacapital4 },
-    ],
-  },
-  {
-    id: '3',
-    name: 'Valstatex',
-    shortDescription: 'Консалтинговая компания Valstatex',
-    logo: '/projects/valstatex/valstatex.svg',
-    description: 'На сайте история и услуги консалтинговой компании Valstatex.',
-    link: 'https://valstatex.uz',
-    stack: 'ReactJS, Ant Design, i18next',
-    photos: [
-      { id: 1, link: valstatex1 },
-      { id: 2, link: valstatex2 },
-      { id: 3, link: valstatex3 },
-    ],
-  },
-  {
-    id: '4',
-    name: 'TruckStaff Solutions',
-    shortDescription: 'Логистическая компания TruckStaff Solutions',
-    logo: '/projects/truckstaff/truckstaff.svg',
-    description: 'Лендинг для американской логистической компании.',
-    link: 'https://truckstaff.us',
-    stack: 'ReactJS, Ant Design, i18next',
-    photos: [
-      { id: 1, link: truckstaff1 },
-      { id: 2, link: truckstaff2 },
-    ],
-  },
-  {
-    id: '5',
-    name: 'Chd',
-    shortDescription: 'Мониторинг учебных центров Chd',
-    logo: '/projects/chd/chd.svg',
-    description: 'Сайт где находится информация о всех учебных центрах города с описанием, ценами и отзывами.',
-    link: '',
-    stack: 'ReactJS, Ant Design, i18next',
-    photos: [
-      { id: 1, link: chd1 },
-      { id: 2, link: chd2 },
-      { id: 3, link: chd3 },
-      { id: 4, link: chd4 },
-    ],
-  },
-  {
-    id: '6',
-    name: 'Edocs',
-    shortDescription: 'Электронный документооборот Edocs',
-    logo: '/projects/edocs/edocs.svg',
-    description: 'Сайт с кабинетом для электронного документооборота в Узбекистане. Я улучшал дизайн, адаптивность и UI проекта.',
-    link: 'https://doc.edocs.uz',
-    stack: 'VueJS, Vuetify',
-    photos: [
-      { id: 1, link: edocs1 },
-      { id: 2, link: edocs2 },
-      { id: 3, link: edocs3 },
-      { id: 4, link: edocs4 },
-      { id: 5, link: edocs5 },
-    ],
-  },
-  {
-    id: '7',
-    name: 'Tech Yard Solutions',
-    shortDescription: 'Аутсорсинговая компания Tech Yard Solutions',
-    logo: '/projects/techyard/techyard.svg',
-    description: 'Сайт аутсорсинговой компании на котором можно было ознакомиться с их услугами. Написан на HTML для улучшения SEO.',
-    link: 'https://techyardsolutions.com',
-    stack: 'HTML, CSS',
-    photos: [
-      { id: 1, link: techyard1 },
-      { id: 2, link: techyard2 },
-      { id: 3, link: techyard3 },
-      { id: 4, link: techyard4 },
-    ],
-  },
-  {
-    id: '8',
-    name: 'IT ACADEMY',
-    shortDescription: 'Учебный центр IT ACADEMY',
-    logo: '/projects/itAcademy/itacademy.svg',
-    description: 'Сайт для IT ACADEMY, где пользователи могли видеть информацию о курсах, почитать блог и познакомиться с историей академии. С нуля до продакшена за 1 месяц. Команда: я и один backend.',
-    link: 'https://it-academy.uz',
-    stack: 'ReactJS, TypeScript, Axios, Ant Design, i18next',
-    photos: [
-      { id: 1, link: itAcademy1 },
-      { id: 2, link: itAcademy2 },
-      { id: 3, link: itAcademy3 },
-      { id: 4, link: itAcademy4 },
-      { id: 5, link: itAcademy5 },
-      { id: 6, link: itAcademy6 },
-      { id: 7, link: itAcademy7 },
-      { id: 8, link: itAcademy8 },
-      { id: 9, link: itAcademy9 },
-      { id: 10, link: itAcademy10 },
-      { id: 11, link: itAcademy11 },
-      { id: 12, link: itAcademy12 },
-    ],
-  },
-  {
-    id: '9',
-    name: 'Kadabra',
-    shortDescription: 'Интернет магазин Kadabra',
-    logo: '/projects/kadabra/kadabra.svg',
-    description: 'Это интернет магазин, где пользователи могли покупать товары или создавать свои магазины и продавать свои товары. С нуля до MVP за 1 месяц. Команда: 2 frontend, 2 backend.',
-    link: '',
-    stack: 'ReactJS, Ant Design, Axios, i18next',
-    photos: [
-      { id: 1, link: kadabra1 },
-      { id: 2, link: kadabra2 },
-      { id: 3, link: kadabra3 },
-      { id: 4, link: kadabra4 },
-      { id: 5, link: kadabra5 },
-      { id: 6, link: kadabra6 },
-      { id: 7, link: kadabra7 },
-      { id: 8, link: kadabra8 },
-      { id: 9, link: kadabra9 },
-    ],
-  },
-  {
-    id: '10',
-    name: 'MATE LMS',
-    shortDescription: 'Учебная платформа MATE LMS',
-    logo: '/projects/mate/mate.svg',
-    description: 'Разрабатывал с нуля удаленную учебную систему (LMS) - MATE Education. Платформа позволяла создавать теорию, тесты и лабораторные работы, а также ввести подробную статистику по каждому студенту, по каждой пройденной им курса/урока/теории/теста/лабораторной вплоть до глав и вопросов. Платформа позволяла студентам получать качественное и актуальное образование, а администраторам контролировать и улучшать учебные курсы. С нуля до продакшена за 4 месяца. Команда: я и один backend разработчик.',
-    link: 'https://mate-edu.io',
-    stack: 'ReactJS, TypeScript, Axios, Ant Design, i18next',
-    photos: [
-      { id: 1, link: mate1 },
-      { id: 2, link: mate2 },
-      { id: 3, link: mate3 },
-      { id: 4, link: mate4 },
-      { id: 5, link: mate5 },
-      { id: 6, link: mate6 },
-      { id: 7, link: mate7 },
-      { id: 8, link: mate8 },
-      { id: 9, link: mate9 },
-      { id: 10, link: mate10 },
-      { id: 11, link: mate11 },
-      { id: 12, link: mate12 },
-      { id: 13, link: mate13 },
-      { id: 14, link: mate14 },
-      { id: 15, link: mate15 },
-      { id: 16, link: mate16 },
-      { id: 17, link: mate17 },
-    ],
-  },
-  {
-    id: '11',
-    name: 'FinAlley',
-    shortDescription: 'Торговля криптовалютой FinAlley',
-    logo: '/projects/finAlley/finalley.svg',
-    description: 'FinAlley - платформа для торговли криптовалютой. Разрабатывал новый функционал, улучшал старый. Также с командой переписывали проекты на TypeScript. Улучшали свою UI библиотеку. В перспективах было закончить разработку платформы которая помогала бы максимально быстро клиентам создовать свои криптобиржы. Команда: 2 frontend + тим.лид, 2 backend, тестировщик, дизайнер, ПМ.',
-    link: 'https://www.finalley.net',
-    stack: 'NextJS, ReactJS, TypeScript, RTK Query, React Bootstrap, Tolgee',
-    photos: [
-      { id: 1, link: finAlley1 },
-      { id: 2, link: finAlley2 },
-      { id: 3, link: finAlley3 },
-      { id: 4, link: finAlley4 },
-      { id: 5, link: finAlley5 },
-      { id: 6, link: finAlley6 },
-      { id: 7, link: finAlley7 },
-      { id: 8, link: finAlley8 },
-      { id: 9, link: finAlley9 },
-      { id: 10, link: finAlley10 },
-      { id: 11, link: finAlley11 },
-      { id: 12, link: finAlley12 },
-    ],
-  },
-  {
-    id: '12',
-    name: 'SkyLab',
-    shortDescription: 'Конструктор соц.сетей SkyLab',
-    logo: '/projects/skylab/skylab.svg',
-    description: 'Проект позволял создавать свои сообщества, свои социальные сети с гибкой настройкой. Клиенту достаточно было настроить соц.сеть в админской панели не прибегая к изменению кода. Команда: 4 frontend, 4 backend, 1 iOS, 1 android, 2 тестировщика, ПМ',
-    link: 'https://skylab.world',
-    stack: 'NextJS, ReactJS, TypeScript, Socket.io, Zustand',
-    photos: [
-      { id: 1, link: skylab1 },
-      { id: 2, link: skylab2 },
-      { id: 3, link: skylab3 },
-      { id: 4, link: skylab4 },
-      { id: 5, link: skylab5 },
-      { id: 6, link: skylab6 },
-      { id: 7, link: skylab7 },
-      { id: 8, link: skylab8 },
-      { id: 9, link: skylab9 },
-      { id: 10, link: skylab10 },
-      { id: 11, link: skylab11 },
-    ],
-  },
-  {
-    id: '13',
-    name: 'Сенат Узбекистана и ООН',
-    shortDescription: 'Платформа мониторинга имплементации стратегий',
-    logo: '/projects/undp/undp.svg',
-    description: 'Проектная работа для сената РУз и ООН. Платформа мониторинга имплементации стратегии достижения гендерного равенства в РУз. Был единственным frontend разработчиком. Разработка с нуля до MVP за 2 месяца. Пока проект приостановлен, идет согласование о продакшене с ООН. Команда: я, один backend разработчик и ПМ он же дизайнер.',
-    link: '',
-    stack: 'Next, Redux-toolkit, TypeScript, Feature sliced design, RTK Query, i18next, Ant Design',
-    photos: [
-      { id: 1, link: undp1 },
-      { id: 2, link: undp2 },
-      { id: 3, link: undp3 },
-      { id: 4, link: undp4 },
-      { id: 5, link: undp5 },
-      { id: 6, link: undp6 },
-      { id: 7, link: undp7 },
-      { id: 8, link: undp8 },
-      { id: 9, link: undp9 },
-      { id: 10, link: undp10 },
-      { id: 11, link: undp11 },
-      { id: 12, link: undp12 },
-      { id: 13, link: undp13 },
-    ],
-  },
-  {
-    id: '14',
-    name: 'RnD платформа',
-    shortDescription: 'RnD платформа',
-    logo: '/projects/rnd/rnd.svg',
-    description: 'RnD платформа для создания проектов, для проектов можно создавать научно исследовательские работы, принимать заявки, открывать вакансии и стажировки, маркетплейс ресурсов (услуги, аренда оборудования, продажа датасетов), создание технических заделов и многое другое. Над разработкой всего этого я принимал участие + внедрение методологии FSD. С MVP до продакшена за 2 месяца. Команда: 2 frontend, тим.лид, тех.лид, 2 backend разработчика, ПМ и дизайнер.',
-    link: '',
-    stack: 'React, Redux-toolkit, RTK Query, TypeScript, Feature sliced design, MUI',
-    photos: [
-      { id: 1, link: rnd1 },
-      { id: 2, link: rnd2 },
-      { id: 3, link: rnd3 },
-      { id: 4, link: rnd4 },
-      { id: 5, link: rnd5 },
-      { id: 6, link: rnd6 },
-      { id: 7, link: rnd7 },
-      { id: 8, link: rnd8 },
-      { id: 9, link: rnd9 },
-      { id: 10, link: rnd10 },
-      { id: 11, link: rnd11 },
-      { id: 12, link: rnd12 },
-      { id: 13, link: rnd13 },
-      { id: 14, link: rnd14 },
-    ],
-  },
-];
