+++ src/components/Preloader/Preloader.tsx
@@ -1,0 +1,7 @@
+import Image from "next/image";
+
+export const Preloader = () => (
+	<div className="bg-black flex h-screen justify-center items-center self-center">
+		<Image src="/spinner.svg" width={200} height={200} alt="Loading" />
+	</div>
+);
