+++ src/components/Preloader/index.ts
@@ -1,0 +1,1 @@
+export { Preloader } from "./Preloader";
