+++ src/page/Skills/index.ts
@@ -1,0 +1,1 @@
+export { Skills } from "./Skills";
