--- tailwind.config.js
@@ -1,24 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-
-module.exports = {
-  content: [
-    './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/components/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
-  ],
-  darkMode: 'class',
-  theme: {
-    letterSpacing: {
-      widest: '.28em',
-    },
-    extend: {
-      screens: {
-        '3xl': '1756px',
-      },
-      fontFamily: {
-        montserrat: ['Montserrat', 'system-ui'],
-      },
-    },
-  },
-  plugins: [],
-};
